+++ net/filter-mirror.c
@@ -47,7 +47,7 @@ static int filter_send(MirrorState *s,
                        int iovcnt)
 {
     NetFilterState *nf = NETFILTER(s);
-    int ret = 0;
+    ssize_t ret = 0;
     size_t size = 0;
     uint32_t len = 0;
     char *buf;
